+++ R/IonDb-methods.R
@@ -52,7 +52,7 @@ setMethod("ions", "IonDb", function(object,
 #' @export
 #'
 #' @rdname IonDb
-setMethod("insertIon", "IonDb", function(object, ions) {
+setMethod("insertIon", "IonDb", function(object, ions, addColumns = FALSE) {
     if (is.data.frame(ions))
         ions$compound_id <- as.character(ions$compound_id)
     .valid_ion(ions, error = TRUE)
@@ -61,14 +61,53 @@ setMethod("insertIon", "IonDb", function(object, ions) {
         if (!all(ions$compound_id %in%
                  dbGetQuery(dbcon, "select compound_id from ms_compound")[, 1]))
             stop("All values of 'compound_id' column of 'ions' must be",
-                 " in 'compound_id' column of 'ms_compound' table of 'cdb'")
-        ions$ion_id <- as.character(
-            seq_len(nrow(ions)) +
-            dbGetQuery(dbcon, "select count(distinct ion_id) from ms_ion")[1,1])
+                 " in 'compound_id' column of the 'ms_compound' table of",
+                 " 'object'. List all avaliable compound IDs with ",
+                 "'compounds(object, c(\"name\", \"compound_id\")'")
+        if (any(colnames(ions) == "ion_id"))
+            warning("Column 'ion_id' will be replaced with internal ",
+                    "identifiers.")
+        max_id <- 0
+        tmp <- dbGetQuery(dbcon, "select max(ion_id) from ms_ion")
+        if (!is.na(tmp[1, 1]))
+            max_id <- as.integer(tmp[1, 1])
+        ions$ion_id <- as.character(max_id + seq_len(nrow(ions)))
+        cols <- colnames(dbGetQuery(dbcon, "select * from ms_ion limit 1"))
+        new_cols <- colnames(ions)[!colnames(ions) %in% cols]
+        if (addColumns && length(new_cols)) {
+            dtype <- dbDataType(dbcon, ions[, new_cols, drop = FALSE])
+            dtype <- paste(names(dtype), dtype)
+            for (dt in dtype)
+                dbExecute(dbcon, paste("alter table ms_ion add", dt))
+            cols <- colnames(dbGetQuery(dbcon, "select * from ms_ion limit 1"))
+            object@.properties$tables$ms_ion <- cols
+        }
         dbAppendTable(dbcon, "ms_ion", ions)
         invisible(object)
     } else stop("Database not initialized")
+    object
 })
+
+#' @importFrom DBI dbGetQuery dbExecute
+#'
+#' @export
+#'
+#' @rdname IonDb
+setMethod("deleteIon", signature(object = "IonDb"),
+          function(object, ids = character(0), ...) {
+              dbcon <- .dbconn(object)
+              if (is.null(dbcon))
+                  stop("Database not initialized")
+              # maybe this check can be removed? 
+              if(any(!ids %in% dbGetQuery(dbcon,
+                                          paste0("select spectrum_id ",
+                                                 "from msms_spectrum"))[, 1]))
+                  warning("Some IDs in 'ids' are not valid and will be ignored")
+              dbExecute(dbcon, paste0("delete from ms_ion where ion_id in (", 
+                                      toString(ids), ")"))
+              object
+          })
+
 
 #' @rdname IonDb
 #'