+++ R/IonDb.R
@@ -4,7 +4,7 @@
 #'
 #' @title IonDb: compound database with additional ion information
 #'
-#' @aliases IonDb-class show,IonDb-method ionVariables ions insertIon
+#' @aliases IonDb-class show,IonDb-method ionVariables ions insertIon deleteIon
 #'
 #' @description
 #'
@@ -65,7 +65,12 @@
 #'
 #' - `insertIon`: allows to add further ions to the `IonDb` object. Note that
 #'   `insertIon` always adds all the ions specified through the `ions` parameter
-#'   and does not check if they are already in the database.
+#'   and does not check if they are already in the database. To add columns
+#'   present in the submitted `data.frame` to the database table set
+#'   `addColumns = TRUE` (default is `addColumns = FALSE`).
+#'   
+#' - `deleteIon`: allows to delete ions from the `IonDb` object by specifying 
+#'    their IDs.
 #'
 #'
 #' @section Filtering the database:
@@ -81,7 +86,11 @@
 #' all supported filters. See also examples below or the usage vignette for
 #' details.
 #'
-#' @param cdb for `IonDb`: `CompDb` object from which data should be
+#' @param addColumns For `insertIons`: `logical(1)` whether columns being
+#'     present in the submitted `data.frame` but not in the database table
+#'     should be added to the database's ion table.
+#'
+#' @param cdb For `IonDb`: `CompDb` object from which data should be
 #'     transferred to the `IonDb` database.
 #'
 #' @param columns For `ions`: `character` with the names of the database
@@ -91,6 +100,10 @@
 #' @param filter For `ions`: filter expression or [AnnotationFilter()] defining
 #'     a filter to be used to retrieve specific elements from the database.
 #'
+#' @param ids For `deleteIon`: `character()` or (alternatively `integer()`) 
+#'     specifying the IDs of the ions to delete. IDs in `ids` that are 
+#'     not associated to any ion in the `IonDb` object are ignored.
+#'     
 #' @param includeId For `ionVariables`: `logical(1)` whether the ion
 #'     ID (column `"ms_ion_id"`) should be included in the result. The
 #'     default is `includeId = FALSE`.
@@ -98,7 +111,8 @@
 #' @param ions for `insertIon` and `IonDb`: `data.frame` with ion definitions
 #'     to be added to the `IonDb` database. Columns `"compound_id"`
 #'     (`character()`), `"ion_adduct"` (`character()`), `"ion_mz"`
-#'     (`numeric()`) and `"ion_rt"` (`numeric()`) are mandatory.
+#'     (`numeric()`) and `"ion_rt"` (`numeric()`) are mandatory (but, with the
+#'     exception of `"compound_id"`, can contain `NA`).
 #'
 #' @param object For all methods: a `IonDb` object.
 #'
@@ -112,6 +126,8 @@
 #'     For all other methods: an `IonDb` object.
 #'
 #' @param ... additional arguments. Currently not used.
+#'
+#' @return See description of the respective function.
 #'
 #' @author Andrea Vicini, Johannes Rainer
 #'
@@ -256,7 +272,7 @@ setValidity("IonDb", function(object) {
 #' @noRd
 .copy_compdb <- function(x, y) {
     tbls <- dbListTables(x)
-    sapply(tbls, function(tbl)
+    lapply(tbls, function(tbl)
         dbWriteTable(y, tbl, dbGetQuery(x, paste0("select * from ", tbl))))
     dbExecute(y, "create index compound_id_idx on ms_compound (compound_id)")
     dbExecute(y, "create index compound_name_idx on ms_compound (name)")